% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoparticles.R
\name{diameter}
\alias{diameter}
\title{Diameter of ZnO nanoparticles from optical band gap}
\usage{
diameter(bandgap, sc = "ZnO")
}
\arguments{
\item{bandgap}{numeric, in eV (vector)
Band gap may not be smaller than the bulk band gap of the material (3.30 eV for ZnO)}

\item{sc}{semiconductor, string (function only supports ZnO)}
}
\value{
diameter, in nanometres (vector)
}
\description{
This formula is simply the inverse of \code{\link{bandgap}}.
Which can be useful to calculate diameter from optically determined band gap.
}
