% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{peak2degrad}
\alias{peak2degrad}
\title{PC degradation by change of abs peak over time}
\usage{
peak2degrad(peak.exp)
}
\arguments{
\item{peak.exp}{dataframe with wavelength and intensity data}
}
\value{
dataframe with the following columns:
   $ sampleid     : chr
   $ minutes      :
   $ idunique     :
   $ n_spectra    :
   $ wl.low       :
   $ wl.high      :
   $ wl.mean      :
   $ abs.mean     :
   $ degrad.ratio :
   $ degrad.log   :
}
\description{
Calculate photocatalytic degradation from the
change of one absorption peak at various time.
}
\details{
peak.exp specification:
   df with wavelength and intensity data for one absorption peak per sample
   with the following REQUIRED columns (yes I know this is an unusual way to do it...)
   >> sampleid
   >> minutes
   >> idunique
   >> n_spectra
   >> wavelength
   >> intensity
}

