% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoparticles.R
\name{bandgap}
\alias{bandgap}
\title{Band gap of ZnO nanoparticles from particle diameter}
\usage{
bandgap(diameter, sc = "ZnO", a, b, c)
}
\arguments{
\item{diameter}{numeric, in nanometre (vector)}

\item{sc}{semiconductor, string (function only supports ZnO)}

\item{a}{parameter a, numeric (optional, default a = 3.940)}

\item{b}{parameter b, numeric (optional, default b = 0.293)}

\item{c}{parameter c, numeric (optional, default c = 3.30)}
}
\value{
bandgap, in eV (vector)
}
\description{
This formula is based on the empirical relation (in turn derived from X-ray
diffraction data, i.e., Scherrer relationship) for ZnO thin-films,
previously published in:
Jacobsson & Edvinsson (2011). Inorganic Chemistry, 50(19), 9578-9586.
https://doi.org/10.1021/ic201327n
}
